+++ libra/src/internal/mod.rs
@@ -1,4 +1,4 @@
 pub mod index;
 pub mod head;
 pub mod branch;
-pub mod protocel;
+pub mod protocol;
